Fixes trac:4180: Move r30225 to 1.7 (silence compiler warnings in test)
---svn-pre-commit-ignore-below---

r28106 [[BR]]
Welcome in the wonderful world of MPI 3.0.

r30018 [[BR]]
Cleanup all the datatype test to avoid any memory leaks or RUI from valgrind.

r30225 [[BR]]
Fix a bunch of compiler warnings in the tests, including:

 * Resolve set-but-not-used issues
 * Resolve incorrect const notation (I checked with George first to see
   what const notation he actually wanted)
 * Comment out unused code (didn't delete it because it's useful
   debugging code)
 * Resolve int<-->void* casting
 * Resolved signed / unsigned comparisons

This commit was SVN r30583.

The following SVN revision numbers were found above:
  r28106 --> open-mpi/ompi@ceb75eae7566e909b0cb78624323eb6c8d2e81fc
  r30018 --> open-mpi/ompi@a85194ae96ec7dc185432a91ab8436dadda7d2d5
  r30225 --> open-mpi/ompi@09f98cb16522a8deda5bf45d1267f8cd789ddc88

The following Trac tickets were found above:
  Ticket 4180 --> https://svn.open-mpi.org/trac/ompi/ticket/4180

+++ test/class/ompi_rb_tree.c
@@ -347,7 +347,7 @@ void test2(void)
         ((ompi_test_rb_value_t *) new_value)->key.bottom = mem[i];
         ((ompi_test_rb_value_t *) new_value)->key.top = 
                                             (void *) ((size_t) mem[i] + size - 1);
-        ((ompi_test_rb_value_t *) new_value)->registered_mpools[0] = (void *) i;
+        ((ompi_test_rb_value_t *) new_value)->registered_mpools[0] = (void *)(intptr_t) i;
         rc = ompi_rb_tree_insert(&tree, &((ompi_test_rb_value_t *)new_value)->key, 
                         new_value);
         if(OMPI_SUCCESS != rc) 
@@ -365,7 +365,7 @@ void test2(void)
         if(NULL == result) 
         {
             test_failure("lookup returned null!");
-        } else if(i != ((int) ((ompi_test_rb_value_t *) result)->registered_mpools[0]))
+        } else if(i != ((int)(intptr_t) ((ompi_test_rb_value_t *) result)->registered_mpools[0]))
         {
             test_failure("lookup returned wrong node!");
         }
@@ -373,7 +373,7 @@ void test2(void)
         if(NULL == result) 
         {
             test_failure("lookup returned null!");
-        } else if(i != ((int) ((ompi_test_rb_value_t *) result)->registered_mpools[0]))
+        } else if(i != ((int)(intptr_t) ((ompi_test_rb_value_t *) result)->registered_mpools[0]))
         {
             test_failure("lookup returned wrong node!");
         }
